requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
typing_extensions==4.12.2
django-widget-tweaks==1.5.0
django-extensions==3.2.3
django-weasyprint==2.4.0
